refactor(toast): ♻️  update toast state & examples
