--- src/new-toast/helpers/index.ts
@@ -1,6 +1,0 @@
-export const genId = (() => {
-  let count = 0;
-  return () => {
-    return (++count).toString();
-  };
-})();
