--- src/new-toast/index.ts
@@ -1,13 +1,0 @@
-export * from "./ToastState";
-export * from "./Toasts";
-export * from "./ToastProvider";
-export * from "./Toasters";
-export * from "./helpers";
-export type {
-  Renderable,
-  ValueFunction,
-  ValueOrFunction,
-  Content,
-  Toast,
-  ToastOptions,
-} from "./ToastTypes";
