+++ src/toast/ToastTypes.ts
@@ -1,9 +1,9 @@
 export type Renderable = JSX.Element | string | number | null;
 
-export type ValueFunction<TValue, TArg> = (arg: TArg) => TValue;
-export type ValueOrFunction<TValue, TArg> =
-  | TValue
-  | ValueFunction<TValue, TArg>;
+export type ValueFunction<Value, Arg> = (arg: Arg) => Value;
+
+export type ValueOrFunction<Value, Arg> = Value | ValueFunction<Value, Arg>;
+
 export type Content = ValueOrFunction<Renderable, Toast>;
 
 export type ToastTypes = "info" | "success" | "warning" | "error";
@@ -19,16 +19,18 @@ export type ToastPlacement =
 export interface Toast {
   createdAt: number;
   id: string;
-  content: Content;
   visible: boolean;
-  reverseOrder: boolean;
   pauseDuration: number;
-  type: ToastTypes;
-  placement: ToastPlacement;
-  duration?: number;
+  content: Content;
   pausedAt?: number;
+  type?: ToastTypes;
+  placement?: ToastPlacement;
+  duration?: number;
+  reverseOrder?: boolean;
 }
 
 export type ToastOptions = Partial<
   Pick<Toast, "id" | "duration" | "type" | "placement" | "reverseOrder">
 >;
+
+export type DefaultToastOptions = Omit<ToastOptions, "id">;
