--- src/toast/__keys.ts
@@ -1,10 +1,0 @@
-// Automatically generated
-const TOAST_STATE_KEYS = [
-  "toasts",
-  "sortedToasts",
-  "showToast",
-  "hideToast",
-  "toggleToast",
-  "removeToast",
-  "isToastVisible",
-] as const;
