+++ src/toast/stories/ToastBasic.stories.tsx
@@ -1,8 +1,14 @@
 import React from "react";
 import { Meta } from "@storybook/react";
 
+import {
+  utilsTemplate,
+  utilsTemplateJs,
+  toastBasicTemplate,
+  toastBasicTemplateJs,
+  toastStyledCssTemplate,
+} from "./templates";
 import { App as Toast } from "./ToastBasic.component";
-import { toastBasicTemplate, toastBasicTemplateJs } from "./templates";
 import { createPreviewTabs } from "../../../scripts/create-preview-tabs";
 
 export default {
@@ -11,9 +17,19 @@ export default {
   parameters: {
     preview: createPreviewTabs({
       js: toastBasicTemplateJs,
+      jsUtils: utilsTemplateJs,
       ts: toastBasicTemplate,
+      tsUtils: utilsTemplate,
+      css: toastStyledCssTemplate,
     }),
   },
+  decorators: [
+    Story => {
+      document.body.id = "tailwind";
+      document.body.classList.add("center");
+      return <Story />;
+    },
+  ],
 } as Meta;
 
 export const Default = () => <Toast />;
