+++ src/toast/stories/ToastReactSpring.component.tsx
@@ -1,17 +1,86 @@
-import React from "react";
+import * as React from "react";
+import { objectKeys } from "@chakra-ui/utils";
 import { animated, useTransition } from "react-spring";
 
-import { Variants, Placements } from "./Utils.component";
-import { ToastProvider, ToastWrapper } from "@renderlesskit/react";
+import {
+  Alert,
+  ToastContainer,
+  TriggerButton,
+  getRandomContent,
+  getRandomType,
+  getPlacementSortedToasts,
+  getRandomPlacement,
+} from "./Utils.component";
+import { ToastProvider, useToasts, useToasters, Toast } from "../index";
 
-const SpringAnimationWrapper: ToastWrapper = ({
-  placement,
-  isVisible,
-  children,
-}) => {
+export const App = () => {
+  return (
+    <ToastProvider duration={5000}>
+      <Notifications />
+      <ToastTriggers />
+    </ToastProvider>
+  );
+};
+
+export default App;
+
+export const Notifications = () => {
+  const { toasts, startPause, endPause, dismissToast } = useToasts();
+
+  const sortedToasts = getPlacementSortedToasts(toasts);
+
+  return (
+    <>
+      {objectKeys(sortedToasts).map(placement => {
+        const toastsList = sortedToasts[placement];
+
+        return (
+          <ToastContainer key={placement} placement={placement}>
+            {toastsList.map(toast => {
+              return (
+                <SpringAnimationWrapper key={toast.id} toast={toast}>
+                  <Alert
+                    toast={toast}
+                    hideToast={dismissToast}
+                    onMouseEnter={() => startPause(toast.id)}
+                    onMouseLeave={() => endPause(toast.id)}
+                  ></Alert>
+                </SpringAnimationWrapper>
+              );
+            })}
+          </ToastContainer>
+        );
+      })}
+    </>
+  );
+};
+
+export function ToastTriggers() {
+  const { showToast } = useToasters();
+
+  return (
+    <div>
+      <TriggerButton
+        onClick={() =>
+          showToast(getRandomContent(), {
+            type: getRandomType(),
+            ...getRandomPlacement(),
+          })
+        }
+      >
+        Add Spring Animated Toast
+      </TriggerButton>
+    </div>
+  );
+}
+
+const SpringAnimationWrapper: React.FC<{ toast: Toast }> = props => {
+  const {
+    toast: { placement = "bottom-right", visible },
+    children,
+  } = props;
   const translate = getTransform(placement, 50);
-
-  const transitions = useTransition(isVisible, null, {
+  const transitions = useTransition(visible, null, {
     from: { opacity: 0, maxHeight: 0, transform: translate.from },
     enter: {
       opacity: 1,
@@ -32,44 +101,6 @@ const SpringAnimationWrapper: ToastWrapper = ({
           ),
       )}
     </>
-  );
-};
-
-export const App: React.FC = () => {
-  return (
-    <ToastProvider
-      autoDismiss={true}
-      placement="bottom-center"
-      animationTimeout={500}
-      toastWrapper={SpringAnimationWrapper}
-      toastTypes={{
-        error: ({ hideToast, content, id }) => {
-          return (
-            <div className="toast" style={{ backgroundColor: "#f02c2d" }}>
-              {content} <button onClick={() => hideToast(id)}>x</button>
-            </div>
-          );
-        },
-        success: ({ hideToast, content, id }) => {
-          return (
-            <div className="toast" style={{ backgroundColor: "#01c24e" }}>
-              {content} <button onClick={() => hideToast(id)}>x</button>
-            </div>
-          );
-        },
-        warning: ({ hideToast, content, id }) => {
-          return (
-            <div className="toast" style={{ backgroundColor: "#ef5013" }}>
-              {content} <button onClick={() => hideToast(id)}>x</button>
-            </div>
-          );
-        },
-      }}
-    >
-      <Variants />
-      <br />
-      <Placements />
-    </ToastProvider>
   );
 };
 
@@ -101,5 +132,3 @@ export const getTransform = (placement: string, pixels: number) => {
 
   return pos;
 };
-
-export default App;
