+++ src/toast/stories/ToastStyled.css
@@ -1,51 +1,9 @@
-.toast {
-  width: 200px;
-  padding: 5px;
-  text-align: center;
-  border-radius: 2px;
-  color: white;
-  transition: 500ms ease-in-out;
-  position: relative;
-}
+@keyframes shrink {
+  from {
+    height: 100%;
+  }
 
-.toast button {
-  width: 25px;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  position: absolute;
-  border: none;
-  outline: none;
-  color: white;
-  cursor: pointer;
-  border-radius: 2px;
-  background-color: rgba(255, 255, 225, 0.4);
-}
-
-.toast button:hover {
-  background-color: rgba(255, 255, 225, 0.2);
-}
-
-.alert-enter {
-  transform: translateY(-50);
-  max-height: 0;
-  opacity: 0;
-}
-
-.alert-enter-active {
-  transform: translateY(0);
-  max-height: 200px;
-  opacity: 1;
-}
-
-.alert-exit {
-  transform: translateX(0) scale(1);
-  max-height: 200px;
-  opacity: 1;
-}
-
-.alert-exit-active {
-  transform: translateX(200px) scale(0.6);
-  max-height: 0;
-  opacity: 0;
+  to {
+    height: 0%;
+  }
 }
